clean up bottom items, move activities icon into its own widget
